test(toolbar): Initial work on testing.
